+++ src/common/js/startup.js
@@ -5,6 +5,6 @@
 // at startup, instanciates a default acAccount instance for the standard 'users' collection
 Meteor.startup(() => {
     if( AccountsCore.configure().autoUsers && !AccountsCore.getInstance( 'users' )){
-        new AccountsCore.acAccount({ name: 'users' });
+        new AccountsCore.Account({ name: 'users' });
     }
 });